Cambios nuevos CargarVenta

+++ WebApiSistemaGestionVentas/Models/Producto.cs
@@ -14,10 +14,10 @@ namespace WebApiSistemaGestionVentas.Models
         public string Descripcion { get; set; }
         public double Costo { get; set; }
         public double PrecioVenta { get; set; }
-        public double Stock { get; set; }
+        public int Stock { get; set; }
         public int IdUsuario { get; set; }
 
-        public Producto(int id, string descripcion, double costo, double precioVenta, double stock, int idUsuario)
+        public Producto(int id, string descripcion, double costo, double precioVenta, int stock, int idUsuario)
         {
             Id = id;
             Descripcion = descripcion;
